requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
psycopg2-binary
PyJWT
sqlalchemy
python-multipart